+++ Userland/Libraries/LibCore/BitStream.h
